Refactor menubar
