+++ src/App.tsx
@@ -3,7 +3,7 @@ import { useGlobalHotkeys, useInitData, useGlobalKeyHandlers } from './hooks';
 
 import BookmarksWindow from './windows/BookmarksWindow';
 import TabsWindow from './windows/TabsWindow';
-import Menubar from './components/Menubar/MenuContainer';
+import Menubar from './components/Menubar/Menubar';
 
 function App() {
   useInitData();
@@ -13,11 +13,8 @@ function App() {
   return (
     <Wrapper>
       <Menubar />
-      
-      <Windows>
-        <TabsWindow />
-        <BookmarksWindow />
-      </Windows>
+      <TabsWindow />
+      <BookmarksWindow />
     </Wrapper>
   );
 }
@@ -25,15 +22,5 @@ function App() {
 export default App;
 
 const Wrapper = styled.div`
-  display: flex;
   height: 600px;
 `;
-
-const Windows = styled.div`
-  flex-grow: 2;
-  height: 100%;
-
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-`
