+++ src/components/Menubar/Menubar.tsx
@@ -1,28 +1,13 @@
 import styled from "styled-components";
-import { MenuItem } from "./MenuContainer";
+import { Apps, Bookmark } from '@styled-icons/material-outlined';
+import MenubarItem from './MenubarItem';
 
-interface Props {
-  menuItems: MenuItem[];
-}
-
-const Menubar = ({ menuItems }: Props) => {
-
-  const MenuItem = ({ item }: { item: MenuItem }) => {
-    const { title, Icon, handleClick } = item;
-
-    return (
-      <ListItem 
-        title={title} 
-        onClick={handleClick} 
-        >
-        <Icon size="24" />
-      </ListItem>
-    );
-  }
+const Menubar = () => {
 
   return (
     <List>
-      { menuItems.map(item => <MenuItem item={item} key={item.title} />) }
+        <MenubarItem title={'탭'} type={'TABS'} Icon={Apps} />
+        <MenubarItem title={'북마크'} type={'BOOKMARKS'} Icon={Bookmark} />
     </List>
   );
 };
@@ -30,19 +15,12 @@ const Menubar = ({ menuItems }: Props) => {
 export default Menubar;
 
 const List = styled.ul`
+  display: flex;
   margin: 0;
   padding: 0;
 
   list-style: none;
   background-image: linear-gradient(120deg, #1d242ed2, #212527);
   color: white;
+  font-size: 0.8rem;
 `;
-
-const ListItem = styled.li`
-  padding: 8px 16px;
-
-  &:hover {
-    cursor: pointer;
-    background-color: #00000086;
-  }
-`
